api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.35
pyjwt==2.9.0
scikit-learn==1.5.2
numpy==2.1.3
